tp4 big share

+++ tp2/val_binaire.c
@@ -6,7 +6,7 @@
 /* Ce code n'est *pas* un exemple à suivre, il ne sert qu'à illuster
    le passage de plus de 8 paramètres à une fonction ! */
 
-uint8_t val_binaire(uint8_t, uint8_t, uint8_t, uint8_t,
+uint16_t val_binaire(uint8_t, uint8_t, uint8_t, uint8_t,
                     uint8_t, uint8_t, uint8_t, uint8_t,
                     uint8_t, uint8_t, uint8_t, uint8_t,
                     uint8_t, uint8_t, uint8_t, uint8_t);
@@ -35,9 +35,9 @@ void stringify(uint8_t s[16])
 int main(void)
 {
    uint8_t toto[16] = {1, 1, 0, 0, 0, 0, 0, 0, 1, 1, 0, 1, 1, 1, 1, 0};
-   uint16_t val = val_binaire( toto[0],  toto[1],  toto[2],  toto[3],
-                               toto[4],  toto[5],  toto[6],  toto[7],
-                               toto[8],  toto[9], toto[10], toto[11],
+   uint16_t val = val_binaire(toto[0],  toto[1],  toto[2],  toto[3],
+                              toto[4],  toto[5],  toto[6],  toto[7],
+                              toto[8],  toto[9],  toto[10], toto[11],
                               toto[12], toto[13], toto[14], toto[15]);
    stringify(toto);
    printf("Valeur du nombre binaire %s = %" PRIx16 "\n", titi, val);
